move more objects to location.py
