+++ lciafmt/location.py
@@ -10,6 +10,9 @@ import json
 
 from esupy.remote import make_url_request
 
+location_meta = ('https://raw.githubusercontent.com/GreenDelta/data/'
+                 'master/refdata/locations.csv')
+
 # %% get GeoJSON
 url = 'https://geography.ecoinvent.org/files'
 
@@ -31,6 +34,17 @@ def extract_coordinates(group) -> dict:
                  for f in features 
                  if f['properties']['shortname'].startswith('US')}
     return us_states
+
+
+def assign_state_names(df):
+    import flowsa
+    f = flowsa.location.get_state_FIPS(abbrev=True).drop(columns='County')
+    f['State'] = f['State'].apply(lambda x: f"US-{x}")
+    fd = f.set_index('FIPS').to_dict()['State']
+    fd['00000'] = 'US'
+    df['Location'] = df['Location'].replace(fd)
+    return df.dropna(subset='Location')
+
 
 if __name__ == "__main__":
     d = extract_coordinates(group='states')
